Metabolite Peak Regulatory Network (GRN) UI and Server Module

This file contains the UI and server logic for the GRN panel. Helper functions for GRN inference and network processing are now located in bulkPanel_GRNFuns.R.

These are the UI and server components of the GRN panel of the shiny app. It is generated by including 'GRN' in the panels.default argument of generateShinyApp.

Usage

BulkGRNUI(id, bulk.metadata, show = TRUE)

BulkGRNServer(id, bulk.intensity.matrix, bulk.metadata, anno)

RegionGRNUI(id, metadata, show = TRUE)

RegionGRNServer(
  id,
  intensity.matrix,
  metadata,
  anno,
  bulk.metadata,
  shared_data
)

Value

The UI and Server components of the shiny module, that can be used within the UI and Server definitions of a shiny app.