Create an MA plot visualising differential analysis (DA) results

This function creates an MA plot to visualise the results of a DA analysis.

bulk_utils_ma_enhance is called indirectly by bulk_utils_ma_plot to add extra features.

Usage

bulk_utils_ma_plot(
  peaks.de.results,
  pval.threshold = 0.05,
  lfc.threshold = 1,
  alpha = 0.1,
  ylims = NULL,
  add.colours = TRUE,
  add.intensity.colour.gradient = TRUE,
  add.guide.lines = TRUE,
  add.labels.auto = TRUE,
  add.labels.custom = FALSE,
  ...
)

bulk_utils_ma_enhance(
  p,
  df,
  pval.threshold,
  lfc.threshold,
  alpha,
  add.colours,
  point.colours = c("#bfbfbf", "orange", "red", "blue"),
  raster = FALSE,
  add.intensity.colour.gradient,
  colour.gradient.scale = list(left = c("#99e6ff", "#000066"), right = c("#99e6ff",
    "#000066")),
  colour.gradient.breaks = ggplot2::waiver(),
  colour.gradient.limits = NULL,
  add.guide.lines,
  guide.line.colours = c("green", "blue"),
  add.labels.auto,
  add.labels.custom,
  annotation = NULL,
  n.labels.auto = c(5, 5, 5),
  peaks.to.label = NULL,
  seed = 0,
  label.force = 1
)

Arguments

peaks.de.results

the table of DA peaks, usually generated by bulk_utils_DA

pval.threshold, lfc.threshold

the p-value and/or log2(fold-change) thresholds to determine whether a peak is DA

alpha

the transparency of points; ignored for DA peaks if add.intensity.colour.gradient is TRUE; default is 0.1

ylims

a single value to create (symmetric) y-axis limits; by default inferred from the data

add.colours

whether to colour peaks based on their log2(fold-change) and -log10(p-value); default is TRUE

add.intensity.colour.gradient

whether to add a colour gradient for DA peaks to present their log2(intensity); default is TRUE

add.guide.lines

whether to add vertical and horizontal guide lines to the plot to highlight the thresholds; default is TRUE

add.labels.auto

whether to automatically label peaks with the highest |log2(fold-change)| and intensity; default is TRUE

add.labels.custom

whether to add labels to user-specified peaks; the parameter peaks.to.label must also be specified; default is FALSE

...

parameters passed on to bulk_utils_ma_enhance

p

MA plot as a ggplot object (usually passed by bulk_utils_ma_plot)

df

data frame of DE results for all peaks (usually passed by bulk_utils_ma_plot)

point.colours

a vector of 4 colours to colour peaks with both pval and lfc under thresholds, just pval under threshold, just lfc under threshold, both pval and lfc over threshold (DE peaks) respectively; only used if add.colours is TRUE

raster

whether to rasterize non-DE peaks with ggraster to reduce memory usage; particularly useful when saving plots to files

colour.gradient.scale

a vector of two colours to create a colour gradient for colouring the DA peaks based on intensity; a named list with components left and right can be supplied to use two different colour scales; only used if add.intensity.colour.gradient is TRUE

colour.gradient.breaks, colour.gradient.limits

parameters to customise the legend of the colour gradient scale; especially useful if creating multiple plots or a plot with two scales; only used if add.intensity.colour.gradient is TRUE

guide.line.colours

a vector with two colours to be used to colour the guide lines; the first colour is used for the p-value and log2(fold-change) thresholds and the second for double those values

annotation

annotation data frame containing a match between the peak field of df (usually ENSEMBL IDs) and the peak names that should be shown in the plot labels; not necessary if df already contains peak names

n.labels.auto

a integer vector of length 3 denoting the number of peaks that should be automatically labelled; the first entry corresponds to DE peaks with the lowest p-value, the second to those with highest absolute log2(fold-change) and the third to those with highest intensity; a single integer can also be specified, to be used for all 3 entries; default is 5

peaks.to.label

a vector of peak names to be labelled in the plot; if names are present those are shown as the labels (but the values are the ones matched - this is to allow custom peak names to be presented)

seed

the random seed to be used for reproducibility; only used for ggrepel::geom_label_repel if labels are present

label.force

passed to the force argument of ggrepel::geom_label_repel; higher values make labels overlap less (at the cost of them being further away from the points they are labelling)

Value

The MA plot as a ggplot object.

The enhanced MA plot as a ggplot object.