Plot a GRN

This function creates a network plot of a GRN.

Usage

bulk_utils_plot_GRN(
  weightMat,
  anno,
  plotConnections,
  plot_position_grid,
  n_networks,
  recurring_regulators
)

Arguments

weightMat

the (weighted) adjacency matrix - regulators in rows, targets in columns

anno

a data frame with peak annotation including 'display_name' (peak names) and 'm_z' (peak IDs); used to highlight target peaks in the network

plotConnections

the number of connections to subset to for plotting

plot_position_grid, n_networks

the position of the plot in the grid (1-4) and the number of networks shown (1-4); these are solely used for hiding unwanted plots in the shiny app

recurring_regulators

targets to be highlighted; usually the result of bulk_utils_find_regulators_with_recurring_edges

Value

A network plot. See visNetwork package for more details.