Plot a GRN

This function creates a network plot of a GRN.

Usage

plot_GRN(
  weightMat,
  anno,
  plotConnections,
  plot_position_grid,
  n_networks,
  recurring_regulators
)

Arguments

weightMat

the (weighted) adjacency matrix - regulators in rows, targets in columns

anno

a data frame with peak annotation including 'display_name' (peak names) and 'm_z' (peak IDs); used to highlight target peaks in the network

plotConnections

the number of connections to subset to for plotting

plot_position_grid, n_networks

the position of the plot in the grid (1-4) and the number of networks shown (1-4); these are solely used for hiding unwanted plots in the shiny app

recurring_regulators

targets to be highlighted; usually the result of find_regulators_with_recurring_edges

Value

A network plot. See visNetwork package for more details.

Examples

weightMat1 <- matrix(
  c(0.1, 0.4, 0.8, 0.3), nrow = 2, ncol = 2,
  dimnames = list("regulators" = c("r1", "r2"), "targets" = c("t1", "t2"))
)
weightMat2 <- matrix(
  c(0.1, 0.2, 0.8, 0.3), nrow = 2, ncol = 2,
  dimnames = list("regulators" = c("r1", "r2"), "targets" = c("t1", "t2"))
)
anno <- tibble::tibble(ENSEMBL = c("r1", "r2", "t1", "t2"), NAME = ENSEMBL)
recurring_regulators <- find_regulators_with_recurring_edges(list(weightMat1, weightMat2), 2)
plot_GRN(weightMat1, anno, 2, 1, 1, recurring_regulators)
#> Warning: Unknown or uninitialised column: `m_z`.

plot_GRN(weightMat2, anno, 2, 1, 1, recurring_regulators)
#> Warning: Unknown or uninitialised column: `m_z`.